double accounts: a body's Energy, TissueJoules and AdultTissueJoules and every booked sum in World are doubles, so the matter residual on the farm reads 1.9e-07 at 3,000 s where the float reserve drifted past 1e-04 (dblA/dblB); WorldState 2, Checkpoint 2; ParallelIdentity re-pinned and Slow

+++ src/Evosim.Core.Bench/StateHash.cs
@@ -53,8 +53,8 @@ namespace Evosim.Core.Bench
                 Organism creature = world.Living[i];
 
                 Long(ref h, creature.Id);
-                Float(ref h, creature.Energy);
-                Float(ref h, creature.TissueJoules);
+                Double(ref h, creature.Energy);
+                Double(ref h, creature.TissueJoules);
                 Float(ref h, creature.Age);
                 Float(ref h, creature.HeightY);
                 Float(ref h, creature.X);
